+++ main/ejml-ddense/generate/org/ejml/dense/block/GeneratorTileMultiplication.java
@@ -23,12 +23,12 @@ import org.ejml.CodeGeneratorMisc;
 import java.io.FileNotFoundException;
 import java.io.PrintStream;
 
-public class GeneratorInnerMultiplication_DDRB {
+public class GeneratorTileMultiplication {
 
     String className;
     PrintStream stream;
 
-    public GeneratorInnerMultiplication_DDRB( String className ) throws FileNotFoundException {
+    public GeneratorTileMultiplication( String className ) throws FileNotFoundException {
         this.className = className;
         stream = new PrintStream(className + ".java");
     }
@@ -62,7 +62,7 @@ public class GeneratorInnerMultiplication_DDRB {
                 "\n" +
                 "/// Matrix multiplication for the inner row major blocks, typically inside [DMatrixRBlock].\n" +
                 "///\n" +
-                "/// This code was auto generated by " + getClass().getSimpleName() + " and should not be modified directly.\n" +
+                "/// Generated by " + getClass().getSimpleName() + ". Do not modify directly.\n" +
                 "@Generated(\"" + getClass().getCanonicalName() + "\")\n" +
                 "public class " + className + " {\n";
 
@@ -283,13 +283,11 @@ public class GeneratorInnerMultiplication_DDRB {
 
         stream.println();
         stream.print(
-                "    /// Performs the follow operation on individual inner blocks:\n" +
-                        "    ///\n" +
                         "    /// C = " + opTypeString + alphaString + "A" + transAString + " \\* B" + transBString + "\n");
     }
 
     private String funcName( Operation opType, boolean transA, boolean transB ) {
-        String name = "blockMult" + opName(opType);
+        String name = "tileMult" + opName(opType);
         if (transA && transB) name += "TransAB";
         else if (transA) name += "TransA";
         else if (transB) name += "TransB";
@@ -348,7 +346,7 @@ public class GeneratorInnerMultiplication_DDRB {
     }
 
     public static void main( String[] args ) throws FileNotFoundException {
-        var app = new GeneratorInnerMultiplication_DDRB("InnerMultiplication_DDRB");
+        var app = new GeneratorTileMultiplication("InnerMultiplication_DDRB");
         app.createClass();
         System.out.println("Done generating class");
     }
