+++ main/ejml-ddense/src/org/ejml/dense/block/TileMultiplication_F64.java
@@ -22,17 +22,15 @@ import javax.annotation.Generated;
 
 /// Matrix multiplication for the inner row major blocks, typically inside [DMatrixRBlock].
 ///
-/// This code was auto generated by GeneratorInnerMultiplication_DDRB and should not be modified directly.
-@Generated("org.ejml.dense.block.GeneratorInnerMultiplication_DDRB")
-public class InnerMultiplication_DDRB {
+/// Generated by GeneratorTileMultiplication. Do not modify directly.
+@Generated("org.ejml.dense.block.GeneratorTileMultiplication")
+public class TileMultiplication_F64 {
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C + A \* B
-    public static void blockMultPlus( final double[] dataA, final double[] dataB, final double[] dataC,
+    public static void tileMultPlus( final double[] dataA, final double[] dataB, final double[] dataC,
                                      final int heightA, final int widthA, final int widthC,
                                      int strideA, int strideB, int strideC,
-                                     int offsetA, int offsetB, int offsetC) {
+                                     int offsetA, int offsetB, int offsetC ) {
 //        for( int i = 0; i < heightA; i++ ) {
 //            for( int k = 0; k < widthA; k++ ) {
 //                for( int j = 0; j < widthC; j++ ) {
@@ -48,11 +46,11 @@ public class InnerMultiplication_DDRB {
             final int rowC = offsetC + i*strideC;
             int k = 0;
             for (; k < kEnd4; k += 4) {
-                double a0 = dataA[aRow + k    ];
+                double a0 = dataA[aRow + k];
                 double a1 = dataA[aRow + k + 1];
                 double a2 = dataA[aRow + k + 2];
                 double a3 = dataA[aRow + k + 3];
-                final int b0 = offsetB + (k    )*strideB;
+                final int b0 = offsetB + (k)*strideB;
                 final int b1 = offsetB + (k + 1)*strideB;
                 final int b2 = offsetB + (k + 2)*strideB;
                 final int b3 = offsetB + (k + 3)*strideB;
@@ -74,22 +72,18 @@ public class InnerMultiplication_DDRB {
         }
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C + A \* B
-    public static void blockMultPlus( final double[] dataA, final double[] dataB, final double[] dataC,
+    public static void tileMultPlus( final double[] dataA, final double[] dataB, final double[] dataC,
                                      final int heightA, final int widthA, final int widthC,
-                                     int offsetA, int offsetB, int offsetC) {
-        blockMultPlus(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
+                                     int offsetA, int offsetB, int offsetC ) {
+        tileMultPlus(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C + A<sup>T</sup> \* B
-    public static void blockMultPlusTransA( final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int strideA, int strideB, int strideC,
-                                     int offsetA, int offsetB, int offsetC) {
+    public static void tileMultPlusTransA( final double[] dataA, final double[] dataB, final double[] dataC,
+                                           final int heightA, final int widthA, final int widthC,
+                                           int strideA, int strideB, int strideC,
+                                           int offsetA, int offsetB, int offsetC ) {
 //        for( int i = 0; i < widthA; i++ ) {
 //            for( int k = 0; k < heightA; k++ ) {
 //                double valA = dataA[k*widthA + i + indexA];
@@ -118,7 +112,7 @@ public class InnerMultiplication_DDRB {
                     bIdx += strideB;
                 }
                 int cIdx = offsetC + i*strideC + j;
-                dataC[cIdx    ] += s0;
+                dataC[cIdx] += s0;
                 dataC[cIdx + 1] += s1;
                 dataC[cIdx + 2] += s2;
                 dataC[cIdx + 3] += s3;
@@ -138,22 +132,18 @@ public class InnerMultiplication_DDRB {
         }
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C + A<sup>T</sup> \* B
-    public static void blockMultPlusTransA( final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int offsetA, int offsetB, int offsetC) {
-        blockMultPlusTransA(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
+    public static void tileMultPlusTransA( final double[] dataA, final double[] dataB, final double[] dataC,
+                                           final int heightA, final int widthA, final int widthC,
+                                           int offsetA, int offsetB, int offsetC ) {
+        tileMultPlusTransA(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C + A \* B<sup>T</sup>
-    public static void blockMultPlusTransB( final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int strideA, int strideB, int strideC,
-                                     int offsetA, int offsetB, int offsetC) {
+    public static void tileMultPlusTransB( final double[] dataA, final double[] dataB, final double[] dataC,
+                                           final int heightA, final int widthA, final int widthC,
+                                           int strideA, int strideB, int strideC,
+                                           int offsetA, int offsetB, int offsetC ) {
 //        for (int i = 0; i < heightA; i++) {
 //            for (int j = 0; j < widthC; j++) {
 //                double val = 0;
@@ -173,7 +163,7 @@ public class InnerMultiplication_DDRB {
             for (; j < jEnd4; j += 4) {
                 double s0 = 0.0, s1 = 0.0, s2 = 0.0, s3 = 0.0;
                 int a = aRow;
-                int b0 = offsetB + (j    )*strideB;
+                int b0 = offsetB + (j)*strideB;
                 int b1 = offsetB + (j + 1)*strideB;
                 int b2 = offsetB + (j + 2)*strideB;
                 int b3 = offsetB + (j + 3)*strideB;
@@ -185,7 +175,7 @@ public class InnerMultiplication_DDRB {
                     s3 += valA*dataB[b3++];
                 }
                 int cIdx = offsetC + i*strideC + j;
-                dataC[cIdx    ] += s0;
+                dataC[cIdx] += s0;
                 dataC[cIdx + 1] += s1;
                 dataC[cIdx + 2] += s2;
                 dataC[cIdx + 3] += s3;
@@ -202,22 +192,18 @@ public class InnerMultiplication_DDRB {
         }
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C + A \* B<sup>T</sup>
-    public static void blockMultPlusTransB( final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int offsetA, int offsetB, int offsetC) {
-        blockMultPlusTransB(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthA, widthC, offsetA, offsetB, offsetC);
+    public static void tileMultPlusTransB( final double[] dataA, final double[] dataB, final double[] dataC,
+                                           final int heightA, final int widthA, final int widthC,
+                                           int offsetA, int offsetB, int offsetC ) {
+        tileMultPlusTransB(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthA, widthC, offsetA, offsetB, offsetC);
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C - A \* B
-    public static void blockMultMinus( final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int strideA, int strideB, int strideC,
-                                     int offsetA, int offsetB, int offsetC) {
+    public static void tileMultMinus( final double[] dataA, final double[] dataB, final double[] dataC,
+                                      final int heightA, final int widthA, final int widthC,
+                                      int strideA, int strideB, int strideC,
+                                      int offsetA, int offsetB, int offsetC ) {
 //        for( int i = 0; i < heightA; i++ ) {
 //            for( int k = 0; k < widthA; k++ ) {
 //                for( int j = 0; j < widthC; j++ ) {
@@ -233,11 +219,11 @@ public class InnerMultiplication_DDRB {
             final int rowC = offsetC + i*strideC;
             int k = 0;
             for (; k < kEnd4; k += 4) {
-                double a0 = dataA[aRow + k    ];
+                double a0 = dataA[aRow + k];
                 double a1 = dataA[aRow + k + 1];
                 double a2 = dataA[aRow + k + 2];
                 double a3 = dataA[aRow + k + 3];
-                final int b0 = offsetB + (k    )*strideB;
+                final int b0 = offsetB + (k)*strideB;
                 final int b1 = offsetB + (k + 1)*strideB;
                 final int b2 = offsetB + (k + 2)*strideB;
                 final int b3 = offsetB + (k + 3)*strideB;
@@ -259,22 +245,18 @@ public class InnerMultiplication_DDRB {
         }
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C - A \* B
-    public static void blockMultMinus( final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int offsetA, int offsetB, int offsetC) {
-        blockMultMinus(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
+    public static void tileMultMinus( final double[] dataA, final double[] dataB, final double[] dataC,
+                                      final int heightA, final int widthA, final int widthC,
+                                      int offsetA, int offsetB, int offsetC ) {
+        tileMultMinus(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C - A<sup>T</sup> \* B
-    public static void blockMultMinusTransA( final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int strideA, int strideB, int strideC,
-                                     int offsetA, int offsetB, int offsetC) {
+    public static void tileMultMinusTransA( final double[] dataA, final double[] dataB, final double[] dataC,
+                                            final int heightA, final int widthA, final int widthC,
+                                            int strideA, int strideB, int strideC,
+                                            int offsetA, int offsetB, int offsetC ) {
 //        for( int i = 0; i < widthA; i++ ) {
 //            for( int k = 0; k < heightA; k++ ) {
 //                double valA = dataA[k*widthA + i + indexA];
@@ -303,7 +285,7 @@ public class InnerMultiplication_DDRB {
                     bIdx += strideB;
                 }
                 int cIdx = offsetC + i*strideC + j;
-                dataC[cIdx    ] -= s0;
+                dataC[cIdx] -= s0;
                 dataC[cIdx + 1] -= s1;
                 dataC[cIdx + 2] -= s2;
                 dataC[cIdx + 3] -= s3;
@@ -323,22 +305,18 @@ public class InnerMultiplication_DDRB {
         }
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C - A<sup>T</sup> \* B
-    public static void blockMultMinusTransA( final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int offsetA, int offsetB, int offsetC) {
-        blockMultMinusTransA(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
+    public static void tileMultMinusTransA( final double[] dataA, final double[] dataB, final double[] dataC,
+                                            final int heightA, final int widthA, final int widthC,
+                                            int offsetA, int offsetB, int offsetC ) {
+        tileMultMinusTransA(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C - A \* B<sup>T</sup>
-    public static void blockMultMinusTransB( final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int strideA, int strideB, int strideC,
-                                     int offsetA, int offsetB, int offsetC) {
+    public static void tileMultMinusTransB( final double[] dataA, final double[] dataB, final double[] dataC,
+                                            final int heightA, final int widthA, final int widthC,
+                                            int strideA, int strideB, int strideC,
+                                            int offsetA, int offsetB, int offsetC ) {
 //        for (int i = 0; i < heightA; i++) {
 //            for (int j = 0; j < widthC; j++) {
 //                double val = 0;
@@ -358,7 +336,7 @@ public class InnerMultiplication_DDRB {
             for (; j < jEnd4; j += 4) {
                 double s0 = 0.0, s1 = 0.0, s2 = 0.0, s3 = 0.0;
                 int a = aRow;
-                int b0 = offsetB + (j    )*strideB;
+                int b0 = offsetB + (j)*strideB;
                 int b1 = offsetB + (j + 1)*strideB;
                 int b2 = offsetB + (j + 2)*strideB;
                 int b3 = offsetB + (j + 3)*strideB;
@@ -370,7 +348,7 @@ public class InnerMultiplication_DDRB {
                     s3 += valA*dataB[b3++];
                 }
                 int cIdx = offsetC + i*strideC + j;
-                dataC[cIdx    ] -= s0;
+                dataC[cIdx] -= s0;
                 dataC[cIdx + 1] -= s1;
                 dataC[cIdx + 2] -= s2;
                 dataC[cIdx + 3] -= s3;
@@ -387,22 +365,18 @@ public class InnerMultiplication_DDRB {
         }
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C - A \* B<sup>T</sup>
-    public static void blockMultMinusTransB( final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int offsetA, int offsetB, int offsetC) {
-        blockMultMinusTransB(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthA, widthC, offsetA, offsetB, offsetC);
+    public static void tileMultMinusTransB( final double[] dataA, final double[] dataB, final double[] dataC,
+                                            final int heightA, final int widthA, final int widthC,
+                                            int offsetA, int offsetB, int offsetC ) {
+        tileMultMinusTransB(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthA, widthC, offsetA, offsetB, offsetC);
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = A \* B
-    public static void blockMultSet( final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int strideA, int strideB, int strideC,
-                                     int offsetA, int offsetB, int offsetC) {
+    public static void tileMultSet( final double[] dataA, final double[] dataB, final double[] dataC,
+                                    final int heightA, final int widthA, final int widthC,
+                                    int strideA, int strideB, int strideC,
+                                    int offsetA, int offsetB, int offsetC ) {
 //        for( int i = 0; i < heightA; i++ ) {
 //            for( int k = 0; k < widthA; k++ ) {
 //                for( int j = 0; j < widthC; j++ ) {
@@ -422,11 +396,11 @@ public class InnerMultiplication_DDRB {
             }
             int k = 0;
             for (; k < kEnd4; k += 4) {
-                double a0 = dataA[aRow + k    ];
+                double a0 = dataA[aRow + k];
                 double a1 = dataA[aRow + k + 1];
                 double a2 = dataA[aRow + k + 2];
                 double a3 = dataA[aRow + k + 3];
-                final int b0 = offsetB + (k    )*strideB;
+                final int b0 = offsetB + (k)*strideB;
                 final int b1 = offsetB + (k + 1)*strideB;
                 final int b2 = offsetB + (k + 2)*strideB;
                 final int b3 = offsetB + (k + 3)*strideB;
@@ -448,22 +422,18 @@ public class InnerMultiplication_DDRB {
         }
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = A \* B
-    public static void blockMultSet( final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int offsetA, int offsetB, int offsetC) {
-        blockMultSet(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
+    public static void tileMultSet( final double[] dataA, final double[] dataB, final double[] dataC,
+                                    final int heightA, final int widthA, final int widthC,
+                                    int offsetA, int offsetB, int offsetC ) {
+        tileMultSet(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = A<sup>T</sup> \* B
-    public static void blockMultSetTransA( final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int strideA, int strideB, int strideC,
-                                     int offsetA, int offsetB, int offsetC) {
+    public static void tileMultSetTransA( final double[] dataA, final double[] dataB, final double[] dataC,
+                                          final int heightA, final int widthA, final int widthC,
+                                          int strideA, int strideB, int strideC,
+                                          int offsetA, int offsetB, int offsetC ) {
 //        for( int i = 0; i < widthA; i++ ) {
 //            for( int k = 0; k < heightA; k++ ) {
 //                double valA = dataA[k*widthA + i + indexA];
@@ -492,7 +462,7 @@ public class InnerMultiplication_DDRB {
                     bIdx += strideB;
                 }
                 int cIdx = offsetC + i*strideC + j;
-                dataC[cIdx    ] = s0;
+                dataC[cIdx] = s0;
                 dataC[cIdx + 1] = s1;
                 dataC[cIdx + 2] = s2;
                 dataC[cIdx + 3] = s3;
@@ -512,22 +482,18 @@ public class InnerMultiplication_DDRB {
         }
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = A<sup>T</sup> \* B
-    public static void blockMultSetTransA( final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int offsetA, int offsetB, int offsetC) {
-        blockMultSetTransA(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
+    public static void tileMultSetTransA( final double[] dataA, final double[] dataB, final double[] dataC,
+                                          final int heightA, final int widthA, final int widthC,
+                                          int offsetA, int offsetB, int offsetC ) {
+        tileMultSetTransA(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = A \* B<sup>T</sup>
-    public static void blockMultSetTransB( final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int strideA, int strideB, int strideC,
-                                     int offsetA, int offsetB, int offsetC) {
+    public static void tileMultSetTransB( final double[] dataA, final double[] dataB, final double[] dataC,
+                                          final int heightA, final int widthA, final int widthC,
+                                          int strideA, int strideB, int strideC,
+                                          int offsetA, int offsetB, int offsetC ) {
 //        for (int i = 0; i < heightA; i++) {
 //            for (int j = 0; j < widthC; j++) {
 //                double val = 0;
@@ -547,7 +513,7 @@ public class InnerMultiplication_DDRB {
             for (; j < jEnd4; j += 4) {
                 double s0 = 0.0, s1 = 0.0, s2 = 0.0, s3 = 0.0;
                 int a = aRow;
-                int b0 = offsetB + (j    )*strideB;
+                int b0 = offsetB + (j)*strideB;
                 int b1 = offsetB + (j + 1)*strideB;
                 int b2 = offsetB + (j + 2)*strideB;
                 int b3 = offsetB + (j + 3)*strideB;
@@ -559,7 +525,7 @@ public class InnerMultiplication_DDRB {
                     s3 += valA*dataB[b3++];
                 }
                 int cIdx = offsetC + i*strideC + j;
-                dataC[cIdx    ] = s0;
+                dataC[cIdx] = s0;
                 dataC[cIdx + 1] = s1;
                 dataC[cIdx + 2] = s2;
                 dataC[cIdx + 3] = s3;
@@ -576,22 +542,18 @@ public class InnerMultiplication_DDRB {
         }
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = A \* B<sup>T</sup>
-    public static void blockMultSetTransB( final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int offsetA, int offsetB, int offsetC) {
-        blockMultSetTransB(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthA, widthC, offsetA, offsetB, offsetC);
+    public static void tileMultSetTransB( final double[] dataA, final double[] dataB, final double[] dataC,
+                                          final int heightA, final int widthA, final int widthC,
+                                          int offsetA, int offsetB, int offsetC ) {
+        tileMultSetTransB(dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthA, widthC, offsetA, offsetB, offsetC);
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C +  &alpha; A \* B
-    public static void blockMultPlus( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
+    public static void tileMultPlus( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
                                      final int heightA, final int widthA, final int widthC,
                                      int strideA, int strideB, int strideC,
-                                     int offsetA, int offsetB, int offsetC) {
+                                     int offsetA, int offsetB, int offsetC ) {
 //        for( int i = 0; i < heightA; i++ ) {
 //            for( int k = 0; k < widthA; k++ ) {
 //                for( int j = 0; j < widthC; j++ ) {
@@ -607,11 +569,11 @@ public class InnerMultiplication_DDRB {
             final int rowC = offsetC + i*strideC;
             int k = 0;
             for (; k < kEnd4; k += 4) {
-                double a0 = alpha*dataA[aRow + k    ];
+                double a0 = alpha*dataA[aRow + k];
                 double a1 = alpha*dataA[aRow + k + 1];
                 double a2 = alpha*dataA[aRow + k + 2];
                 double a3 = alpha*dataA[aRow + k + 3];
-                final int b0 = offsetB + (k    )*strideB;
+                final int b0 = offsetB + (k)*strideB;
                 final int b1 = offsetB + (k + 1)*strideB;
                 final int b2 = offsetB + (k + 2)*strideB;
                 final int b3 = offsetB + (k + 3)*strideB;
@@ -633,22 +595,18 @@ public class InnerMultiplication_DDRB {
         }
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C +  &alpha; A \* B
-    public static void blockMultPlus( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
+    public static void tileMultPlus( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
                                      final int heightA, final int widthA, final int widthC,
-                                     int offsetA, int offsetB, int offsetC) {
-        blockMultPlus(alpha, dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
+                                     int offsetA, int offsetB, int offsetC ) {
+        tileMultPlus(alpha, dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C +  &alpha; A<sup>T</sup> \* B
-    public static void blockMultPlusTransA( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int strideA, int strideB, int strideC,
-                                     int offsetA, int offsetB, int offsetC) {
+    public static void tileMultPlusTransA( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
+                                           final int heightA, final int widthA, final int widthC,
+                                           int strideA, int strideB, int strideC,
+                                           int offsetA, int offsetB, int offsetC ) {
 //        for( int i = 0; i < widthA; i++ ) {
 //            for( int k = 0; k < heightA; k++ ) {
 //                double valA = alpha*dataA[k*widthA + i + indexA];
@@ -677,7 +635,7 @@ public class InnerMultiplication_DDRB {
                     bIdx += strideB;
                 }
                 int cIdx = offsetC + i*strideC + j;
-                dataC[cIdx    ] += s0;
+                dataC[cIdx] += s0;
                 dataC[cIdx + 1] += s1;
                 dataC[cIdx + 2] += s2;
                 dataC[cIdx + 3] += s3;
@@ -697,22 +655,18 @@ public class InnerMultiplication_DDRB {
         }
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C +  &alpha; A<sup>T</sup> \* B
-    public static void blockMultPlusTransA( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int offsetA, int offsetB, int offsetC) {
-        blockMultPlusTransA(alpha, dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
+    public static void tileMultPlusTransA( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
+                                           final int heightA, final int widthA, final int widthC,
+                                           int offsetA, int offsetB, int offsetC ) {
+        tileMultPlusTransA(alpha, dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C +  &alpha; A \* B<sup>T</sup>
-    public static void blockMultPlusTransB( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int strideA, int strideB, int strideC,
-                                     int offsetA, int offsetB, int offsetC) {
+    public static void tileMultPlusTransB( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
+                                           final int heightA, final int widthA, final int widthC,
+                                           int strideA, int strideB, int strideC,
+                                           int offsetA, int offsetB, int offsetC ) {
 //        for (int i = 0; i < heightA; i++) {
 //            for (int j = 0; j < widthC; j++) {
 //                double val = 0;
@@ -732,7 +686,7 @@ public class InnerMultiplication_DDRB {
             for (; j < jEnd4; j += 4) {
                 double s0 = 0.0, s1 = 0.0, s2 = 0.0, s3 = 0.0;
                 int a = aRow;
-                int b0 = offsetB + (j    )*strideB;
+                int b0 = offsetB + (j)*strideB;
                 int b1 = offsetB + (j + 1)*strideB;
                 int b2 = offsetB + (j + 2)*strideB;
                 int b3 = offsetB + (j + 3)*strideB;
@@ -744,7 +698,7 @@ public class InnerMultiplication_DDRB {
                     s3 += valA*dataB[b3++];
                 }
                 int cIdx = offsetC + i*strideC + j;
-                dataC[cIdx    ] += alpha*s0;
+                dataC[cIdx] += alpha*s0;
                 dataC[cIdx + 1] += alpha*s1;
                 dataC[cIdx + 2] += alpha*s2;
                 dataC[cIdx + 3] += alpha*s3;
@@ -761,22 +715,18 @@ public class InnerMultiplication_DDRB {
         }
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C = C +  &alpha; A \* B<sup>T</sup>
-    public static void blockMultPlusTransB( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int offsetA, int offsetB, int offsetC) {
-        blockMultPlusTransB(alpha, dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthA, widthC, offsetA, offsetB, offsetC);
+    public static void tileMultPlusTransB( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
+                                           final int heightA, final int widthA, final int widthC,
+                                           int offsetA, int offsetB, int offsetC ) {
+        tileMultPlusTransB(alpha, dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthA, widthC, offsetA, offsetB, offsetC);
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C =  &alpha; A \* B
-    public static void blockMultSet( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int strideA, int strideB, int strideC,
-                                     int offsetA, int offsetB, int offsetC) {
+    public static void tileMultSet( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
+                                    final int heightA, final int widthA, final int widthC,
+                                    int strideA, int strideB, int strideC,
+                                    int offsetA, int offsetB, int offsetC ) {
 //        for( int i = 0; i < heightA; i++ ) {
 //            for( int k = 0; k < widthA; k++ ) {
 //                for( int j = 0; j < widthC; j++ ) {
@@ -796,11 +746,11 @@ public class InnerMultiplication_DDRB {
             }
             int k = 0;
             for (; k < kEnd4; k += 4) {
-                double a0 = alpha*dataA[aRow + k    ];
+                double a0 = alpha*dataA[aRow + k];
                 double a1 = alpha*dataA[aRow + k + 1];
                 double a2 = alpha*dataA[aRow + k + 2];
                 double a3 = alpha*dataA[aRow + k + 3];
-                final int b0 = offsetB + (k    )*strideB;
+                final int b0 = offsetB + (k)*strideB;
                 final int b1 = offsetB + (k + 1)*strideB;
                 final int b2 = offsetB + (k + 2)*strideB;
                 final int b3 = offsetB + (k + 3)*strideB;
@@ -822,22 +772,18 @@ public class InnerMultiplication_DDRB {
         }
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C =  &alpha; A \* B
-    public static void blockMultSet( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int offsetA, int offsetB, int offsetC) {
-        blockMultSet(alpha, dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
+    public static void tileMultSet( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
+                                    final int heightA, final int widthA, final int widthC,
+                                    int offsetA, int offsetB, int offsetC ) {
+        tileMultSet(alpha, dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C =  &alpha; A<sup>T</sup> \* B
-    public static void blockMultSetTransA( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int strideA, int strideB, int strideC,
-                                     int offsetA, int offsetB, int offsetC) {
+    public static void tileMultSetTransA( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
+                                          final int heightA, final int widthA, final int widthC,
+                                          int strideA, int strideB, int strideC,
+                                          int offsetA, int offsetB, int offsetC ) {
 //        for( int i = 0; i < widthA; i++ ) {
 //            for( int k = 0; k < heightA; k++ ) {
 //                double valA = alpha*dataA[k*widthA + i + indexA];
@@ -866,7 +812,7 @@ public class InnerMultiplication_DDRB {
                     bIdx += strideB;
                 }
                 int cIdx = offsetC + i*strideC + j;
-                dataC[cIdx    ] = s0;
+                dataC[cIdx] = s0;
                 dataC[cIdx + 1] = s1;
                 dataC[cIdx + 2] = s2;
                 dataC[cIdx + 3] = s3;
@@ -886,22 +832,18 @@ public class InnerMultiplication_DDRB {
         }
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C =  &alpha; A<sup>T</sup> \* B
-    public static void blockMultSetTransA( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int offsetA, int offsetB, int offsetC) {
-        blockMultSetTransA(alpha, dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
+    public static void tileMultSetTransA( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
+                                          final int heightA, final int widthA, final int widthC,
+                                          int offsetA, int offsetB, int offsetC ) {
+        tileMultSetTransA(alpha, dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthC, widthC, offsetA, offsetB, offsetC);
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C =  &alpha; A \* B<sup>T</sup>
-    public static void blockMultSetTransB( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int strideA, int strideB, int strideC,
-                                     int offsetA, int offsetB, int offsetC) {
+    public static void tileMultSetTransB( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
+                                          final int heightA, final int widthA, final int widthC,
+                                          int strideA, int strideB, int strideC,
+                                          int offsetA, int offsetB, int offsetC ) {
 //        for (int i = 0; i < heightA; i++) {
 //            for (int j = 0; j < widthC; j++) {
 //                double val = 0;
@@ -921,7 +863,7 @@ public class InnerMultiplication_DDRB {
             for (; j < jEnd4; j += 4) {
                 double s0 = 0.0, s1 = 0.0, s2 = 0.0, s3 = 0.0;
                 int a = aRow;
-                int b0 = offsetB + (j    )*strideB;
+                int b0 = offsetB + (j)*strideB;
                 int b1 = offsetB + (j + 1)*strideB;
                 int b2 = offsetB + (j + 2)*strideB;
                 int b3 = offsetB + (j + 3)*strideB;
@@ -933,7 +875,7 @@ public class InnerMultiplication_DDRB {
                     s3 += valA*dataB[b3++];
                 }
                 int cIdx = offsetC + i*strideC + j;
-                dataC[cIdx    ] = alpha*s0;
+                dataC[cIdx] = alpha*s0;
                 dataC[cIdx + 1] = alpha*s1;
                 dataC[cIdx + 2] = alpha*s2;
                 dataC[cIdx + 3] = alpha*s3;
@@ -950,12 +892,10 @@ public class InnerMultiplication_DDRB {
         }
     }
 
-    /// Performs the follow operation on individual inner blocks:
-    ///
     /// C =  &alpha; A \* B<sup>T</sup>
-    public static void blockMultSetTransB( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
-                                     final int heightA, final int widthA, final int widthC,
-                                     int offsetA, int offsetB, int offsetC) {
-        blockMultSetTransB(alpha, dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthA, widthC, offsetA, offsetB, offsetC);
+    public static void tileMultSetTransB( double alpha, final double[] dataA, final double[] dataB, final double[] dataC,
+                                          final int heightA, final int widthA, final int widthC,
+                                          int offsetA, int offsetB, int offsetC ) {
+        tileMultSetTransB(alpha, dataA, dataB, dataC, heightA, widthA, widthC, widthA, widthA, widthC, offsetA, offsetB, offsetC);
     }
 }
